requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
python-multipart==0.0.6
pillow==10.0.1
torch==2.0.1nv
transformers==4.33.2
datasets==2.14.5
tensorboard==2.14.0
tqdm==4.66.1
pandas==2.1.1
jinja2==3.1.2
aiofiles==23.2.1
python-dotenv==1.0.0
matplotlib==3.8.0
numpy==1.26.0
nltk==3.8.1
evaluate==0.4.1
sentencepiece==0.1.99